requirements.txt
customtkinter==4.6.0
altgraph==0.17.2
attrs==21.4.0
auto-py-to-exe==2.18.0
Automat==20.2.0
Babel==2.9.1
beautifulsoup4==4.10.0
bottle==0.12.19
bottle-websocket==0.2.9
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorzero==2.0
constantly==15.1.0
cryptocode==0.1
datefinder==0.7.1
dill==0.3.4
docopt==0.6.2
Eel==0.12.4
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
gpiozero==1.6.2
GPUtil==1.4.0
greenlet==1.1.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
MouseInfo==0.1.3
mss==6.1.0
multiprocess==0.70.12.2
names==0.3.0
numpy==1.21.4
opencv-python==4.5.4.60
pefile==2021.9.3
Pillow==9.0.1
pipreqs==0.4.11
playsound==1.2.2
psutil==5.9.0
py-cpuinfo==8.0.0
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodomex==3.14.1
pyflakes==2.4.0
pygame==2.1.2
PyGetWindow==0.0.9
pyglet==1.5.22
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.2
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PyRect==0.1.4
PyScreeze==0.1.28
PySide2==5.15.2.1
python-dateutil==2.8.2
python-qt5==0.1.10
pytweening==1.0.4
pytz==2021.3
pywin32==303
pywin32-ctypes==0.2.0
PyYAML==6.0
qt4reactor==1.6
regex==2022.3.2
requests==2.27.1
screeninfo==0.8
shiboken2==5.15.2.1
six==1.16.0
soupsieve==2.3.1
tkfilebrowser==2.3.2
Twisted==22.1.0
twisted-iocpsupport==1.0.2
typing_extensions==4.1.1
urllib3==1.26.8
websockets==10.2
wget==3.2
whichcraft==0.6.1
WMI==1.5.1
yarg==0.1.9
zope.event==4.5.0
zope.interface==5.4.0
